Refactored: Moved top and bottom bar into Library and refactored layout
