+++ src/components/library/Grid.less
@@ -1,0 +1,53 @@
+
+@scrollbarWidth: 20px;  // In fact it's probably less, but we use a little extra space
+
+
+.grid-cell-size(@outerWidth; @columnCount) {
+    grid-template-columns: repeat(@columnCount, @gridCellSize);
+    width: (@columnCount * @gridCellSize);
+}
+
+.grid-cell-media-between(@outerWidth; @columnCount) {
+    @media (max-width: ((@columnCount + 1) * @gridCellSize + @outerWidth - 1px)) and (min-width: (@columnCount * @gridCellSize + @outerWidth)) {
+        .grid-cell-size(@outerWidth; @columnCount);
+    }
+}
+
+.grid-cell-media-last(@outerWidth; @columnCount) {
+  @media (min-width: (@columnCount * @gridCellSize + @outerWidth)) {
+      .grid-cell-size(@outerWidth; @columnCount);
+  }
+}
+
+
+.Grid {
+    display: grid;
+    grid-template-rows: auto;
+    justify-items: center;
+    margin: auto;
+
+    .grid-cell-size(@scrollbarWidth; 1);
+    .grid-cell-media-between(@scrollbarWidth; 2);
+    .grid-cell-media-between(@scrollbarWidth; 3);
+    .grid-cell-media-between(@scrollbarWidth; 4);
+    .grid-cell-media-between(@scrollbarWidth; 5);
+    .grid-cell-media-between(@scrollbarWidth; 6);
+    .grid-cell-media-between(@scrollbarWidth; 7);
+    .grid-cell-media-between(@scrollbarWidth; 8);
+    .grid-cell-media-between(@scrollbarWidth; 9);
+    .grid-cell-media-last(@scrollbarWidth; 10);
+
+
+    .Ansel.hasSidebar & {
+        .grid-cell-size(@sidebarWidth + @scrollbarWidth; 1);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 2);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 3);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 4);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 5);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 6);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 7);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 8);
+        .grid-cell-media-between(@sidebarWidth + @scrollbarWidth; 9);
+        .grid-cell-media-last(@sidebarWidth + @scrollbarWidth; 10);
+    }
+}
