+++ src/components/library/LibraryBottomBar.less
@@ -1,0 +1,21 @@
+.LibraryBottomBar-content {
+    position: relative;
+    line-height: 48px;
+    margin: 0 25px;
+}
+
+.LibraryBottomBar-selection {
+    background: @blue-grey-900;
+    border-radius: 5px;
+    padding: 4px 10px;
+}
+
+.LibraryBottomBar-deselectAll:before {
+    content: "|";
+    padding-right: 6px;
+}
+
+.LibraryBottomBar-total {
+    position: absolute;
+    left: 48%;
+}
