+++ src/components/main/Container.less
@@ -1,0 +1,9 @@
+.Container {
+    background: @blue-grey-100;
+    color: @white;
+}
+
+.Container-body {
+    width: 100%;
+    height: 100%;
+}
