+++ src/less/components.less
@@ -1,5 +1,3 @@
-@import "./header.less";
-@import './grid.less';
 @import "./sidebar.less";
 @import "./splash.less";
 
@@ -7,6 +5,12 @@
 @import "../components/detail/PictureDetail.less";
 @import "../components/detail/PictureInfo.less";
 @import "../components/main/Ansel.less";
+@import "../components/main/Container.less";
+@import "../components/library/Grid.less";
+@import "../components/library/Library.less";
+@import "../components/library/LibraryBottomBar.less";
+@import "../components/library/LibraryTopBar.less";
+@import "../components/library/Picture.less";
 @import "../components/widget/Button.less";
 @import "../components/widget/ButtonGroup.less";
 @import "../components/widget/Toolbar.less";
