--- src/less/header.less
@@ -1,34 +1,0 @@
-/*Header*/
-#top-bar {
-  width: ~"calc(100% - 300px)";
-  border-left: 1px solid @blue-grey-900;
-  position: relative;
-  float: right;
-  transition: width @transition-time ease-in;
-
-  button {
-    font-size: 2em;
-    padding: 1px 12px;
-  }
-
-  button.flagged.active {
-    color: @red;
-  }
-
-  button#hamburger {
-    opacity: 0;
-    height: 0;
-    transition: opacity @transition-time ease-out;
-  }
-}
-
-#top-bar.no-sidebar { 
-  width: 100%;
-  transition: width @transition-time ease-out;
-
-  button#hamburger {
-    opacity: 1;
-    height: auto;
-    transition: opacity @transition-time ease-in;
-  }
-}
