+++ src/less/index.less
@@ -8,11 +8,13 @@
   box-sizing: border-box;
 }
 
-html, body, #app, #ansel, #sidebar, #container {
+html, body, #app, .Ansel {
+  width: 100%;
   height: 100%;
 }
 body {
   margin: 0;
+  overflow: hidden;
 }
 
 body, button {
@@ -54,21 +56,6 @@ svg.logo {
   .st3{opacity:0.6;fill:#E72174;}
   .st4{opacity:0.6;fill:#253887;}
   .st5{opacity:0.6;fill:#3BA9E0;}
-}
-
-.card {
-  display: flex;
-  flex-direction: column;
-  font-size: 16px;
-  font-weight: 400;
-  min-height: 200px;
-  overflow: hidden;
-  width: 330px;
-  z-index: 1;
-  position: relative;
-  background: rgb(255,255,255);
-  border-radius: 2px;
-  box-sizing: border-box;
 }
 
 input.textfield {
@@ -143,15 +130,15 @@ button {
 div.outer-modal {
   position: fixed;
   top: 0;
-  left: 300px;
-  width: ~"calc(100% - 300px)";
+  left: @sidebarWidth;
+  width: ~"calc(100% - @{sidebarWidth})";
   height: 100%;
   background-color: rgba(0,0,0,0.6);
   z-index: 3;
 
   div.modal {
     width: 400px;
-    height: 300px;
+    height: @sidebarWidth;
     background-color: @blue-grey-100;
     color: @blue-grey-900;
     position: absolute;
@@ -162,20 +149,6 @@ div.outer-modal {
   }
 }
 
-#container {
-  position: relative;
-  z-index: 1;
-  float: right;
-  width: ~"calc(100% - 300px)";
-  height: ~"calc(100% - 48px)";
-  background: @blue-grey-100;
-  color: @white;
-  overflow-y: auto;
-  transition: width @transition-time ease-in;
-
-  #library {
-    height: 100%;
-
     /* create a full-height inline block pseudo=element */
     div.picture-diff div.before:before, 
     div.picture-diff div.after:before { 
@@ -183,13 +156,6 @@ div.outer-modal {
       display: inline-block;
       vertical-align: middle;  /* vertical alignment of the inline element */
       height: 100%;
-    }
-
-    .Library-body {
-        width: 100%;
-        height: 100%;
-        margin: 0 auto;
-        position: relative;
     }
 
     div.picture-diff {
@@ -247,41 +213,6 @@ div.outer-modal {
       }
     }
 
-    a.picture {
-      float: left;
-      margin: 5px;
-      padding: 5px;
-      background-color: transparent;
-      width: 247px;
-      height: 247px;
-      display: block;
-      text-align: center;
-      transition: background-color 0.2s ease-in;
-
-      span.v-align {
-        display: inline-block;
-        height: 100%;
-        vertical-align: middle;
-      }
-
-      img {
-        max-width: 237px;
-        vertical-align: middle;
-        max-height: 240px;
-      }
-    }
-
-    a.picture.highlighted {
-      background: @deep-orange-500;
-    }
-  }
-}
-
-#container.no-sidebar { 
-  width: 100%;
-  transition: width @transition-time ease-out;
-}
-
 #progress {
   color: @blue-grey-900;
   text-align: center;
@@ -302,7 +233,7 @@ div.outer-modal {
   }
 }
 
-#container div.settings-container {
+div.settings-container {
   background: @blue-grey-800;
   height: 100%;
   padding: 0 42px;
