+++ src/less/variables.less
@@ -18,3 +18,8 @@
 @white: #ffffff;
 
 @transition-time: 0.4s;
+
+@sidebarWidth: 300px;
+@thumbnailSize: 250px;
+@gridCellPadding: 5px;
+@gridCellSize: @thumbnailSize + 2 * @gridCellPadding;
